+++ example/AppStyleX/src/Circle/index.js
@@ -37,8 +37,8 @@ const Circle = () => {
   };
 
   useEffect(() => {
-    console.log(' --- xdebug isDarkTheme', isDarkTheme, isDarkTheme ? 1 : 0);
     timing(animatedCircle.current, {
+      useNativeDriver: false,
       toValue: isDarkTheme ? 1 : 0,
       duration: 250,
     }).start();