+++ example/AppStyleX/src/ToggleButton/index.js
@@ -6,14 +6,15 @@ import useIsDark from '../style/hooks/useIsDark';
 
 import useStyles, {ROOT_WIDTH} from './styles';
 
-
 const {Value, spring} = Animated;
 const HALF_ROOT_WIDTH = ROOT_WIDTH / 2;
 
 const ToggleButton = ({onPress}) => {
   const styles = useStyles();
   const isDark = useIsDark();
-  const bgStyle = {backgroundColor: useColorTransition(({palette}) => palette.bg)};
+  const bgStyle = {
+    backgroundColor: useColorTransition(({palette}) => palette.bg),
+  };
   const animatedOffset = useRef(new Value(0));
   const animatedStyle = {
     transform: [
@@ -28,6 +29,7 @@ const ToggleButton = ({onPress}) => {
 
   useEffect(() => {
     spring(animatedOffset.current, {
+      useNativeDriver: false,
       toValue: isDark ? 1 : 0,
       duration: 250,
     }).start();