+++ src/__tests__/dependencyUsage.test.ts
@@ -1,0 +1,13 @@
+import { onUse, resetUsing, getUsing } from "../dependencyUsage";
+
+it("should clear usage map", () => {
+  onUse("a");
+  onUse("a");
+  onUse("b");
+
+  expect(getUsing()).toEqual({ a: true, b: true });
+
+  resetUsing();
+
+  expect(getUsing()).toEqual({});
+});