+++ src/__tests__/dimensions.test.ts
@@ -1,0 +1,36 @@
+import {
+  WINDOW_DEPENDENCY_KEY,
+  SCREEN_DEPENDENCY_KEY
+} from "../dimensions/consts";
+import { optimizeDependencies } from "../dimensions/index";
+
+describe("optimizeDependencies", () => {
+  it("should remove screen dependency when window dependency exist", () => {
+    expect(
+      optimizeDependencies({
+        [WINDOW_DEPENDENCY_KEY]: true,
+        [SCREEN_DEPENDENCY_KEY]: true
+      })
+    ).toEqual({
+      [WINDOW_DEPENDENCY_KEY]: true
+    });
+  });
+
+  it("should skip removing", () => {
+    expect(
+      optimizeDependencies({
+        [WINDOW_DEPENDENCY_KEY]: true
+      })
+    ).toEqual({
+      [WINDOW_DEPENDENCY_KEY]: true
+    });
+
+    expect(
+      optimizeDependencies({
+        [SCREEN_DEPENDENCY_KEY]: true
+      })
+    ).toEqual({
+      [SCREEN_DEPENDENCY_KEY]: true
+    });
+  });
+});