+++ src/index.ts
@@ -1,12 +1,5 @@
+export { useTheme } from "./useTheme";
+export { withStyles } from "./withStyles";
 export { ThemeProvider, ThemeConsumer } from "./context";
-export { default as useTheme } from "./useTheme";
-export { default as withStyles } from "./withStyles";
-export { default as makeUseStyles, MakeUseStylesFn } from "./makeUseStyles";
-export { default as useColorTransition } from "./useColorTransition";
-export {
-  maxWidth,
-  maxHeight,
-  minHeight,
-  minWidth,
-  createDimensionQueryHelper
-} from "./mediaQueryHelpers";
+export { useColorTransition } from "./useColorTransition";
+export { makeUseStyles, MakeUseStylesFn } from "./makeUseStyles";