+++ src/safe-area/eventEmitter.ts
@@ -1,0 +1,4 @@
+import { createEventEmitter } from "../createEventEmitter";
+import { SAFE_AREA_DEPENDENCY_KEY } from "./consts";
+
+export const { emit, on } = createEventEmitter(SAFE_AREA_DEPENDENCY_KEY);