+++ src/safe-area/init.tsx
@@ -1,0 +1,8 @@
+import * as React from "react";
+
+import { addDependency } from "../dependencyRegistry";
+
+import { SAFE_AREA_DEPENDENCY_KEY } from "./consts";
+import { on } from "./eventEmitter";
+
+addDependency(SAFE_AREA_DEPENDENCY_KEY, (handler: () => void) => on(handler));