+++ src/useTheme.ts
@@ -3,6 +3,4 @@ import { useContext } from "react";
 import { themeContext } from "./context";
 
 // @ts-ignore
-const useTheme = <T>(): T => useContext(themeContext);
-
-export default useTheme;
+export const useTheme = <T>(): T => useContext(themeContext);